refactor(cli): migrate the validate-only atmosphere backends onto discovery
Stand up earthlens._atmosphere_cli and its earthlens.cli entry point, and move the
offline structural validators for catrare / drought / goes / mswep / nrel / pvgis /
radklim into earthlens.<backend>.cli. Relocate their unit tests; the bundled-clean
parametrised checks keep covering them via the merged registry. Shrink the guard
allow-list (seven backends).
